start decoupling LAN/netplay dialogs from the actual modules
